bot/pyproject.toml
[tool.poetry]
name = "street-bot-server"
version = "0.1.0"
description = ""
authors = ["chezbgone <chezbgone@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.4.0"
common = {path = "../common"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
